Changed assetmodeldelete

+++ Backend/controllers/assetModelController.js
@@ -1,4 +1,5 @@
 import AssetModel from "../models/Asset.js";
+import AssetItem from "../models/AssetItem.js";
 
 export const addAssetModel = async (req, res) => {
     try {
@@ -42,16 +43,22 @@ export const editAssetModel = async (req, res) => {
 
 export const deleteAssetModel = async (req, res) => {
     try {
+         let { id } = req.query;
+        let assetItem = await AssetItem.findOne({model:id})
 
-        let { id } = req.query;
-        let response= await AssetModel.findByIdAndDelete(id)
-        if(response){
-             return res.status(200).send({ message: "AssetModel Deleted" });
+        if (assetItem) {
+            return res.status(400).send({ error: "Asset Items are present please delete all the asset Items" })
         }
-       else {
-            return res.status(400).send({error:"Asset Model Not found"})
-       }
+        else {
+            let response = await AssetModel.findByIdAndDelete(id)
+            if (response) {
+                return res.status(200).send({ message: "AssetModel Deleted" });
+            }
+            else {
+                return res.status(400).send({ error: "Asset Model Not found" })
+            }
 
+        }
     } catch (error) {
         return res.status(500).send({
             message: "Something went wrong",
@@ -81,3 +88,23 @@ export const getAllAssetModels = async (req, res) => {
         });
     }
 };
+
+export const getAllAssetModelsWithItems = async (req, res) => {
+    try {
+        let allModelsWithItems = await AssetModel.aggregate([{
+            $lookup: {
+                localField: "_id",
+                from: "assetitems",
+                foreignField: "model",
+                as: "items",
+            }
+        }])
+        res.status(200).send(allModelsWithItems)
+
+    } catch (error) {
+        return res.status(500).send({
+            message: "Something went wrong",
+            error: error.message
+        });
+    }
+}
